Feature selection for training set
the code is not clean right now, but it has the code for feature extraction during user interaction. A screenshot will be recorded every 2 seconds and the pixel array will be stored as the feature input. Steering value [-1,1] will be appended to position 0 as the label. there are 20 labels between [-1.0, -0.9, ... , 0.9, 1.0] this will be used later as the size of the NN output layer